CIMPartition: fix a clang-tidy narrowing-conversion failure in static-analysis CI
lib/Transforms/CIMPartition.cpp:160-161 (the ragged-shape zero-padding
from an earlier commit) assigned llvm::divideCeil's uint64_t return
straight into an int64_t local. bugprone-narrowing-conversions correctly
flags that as implementation-defined in general, and the project's
static-analysis CI job runs clang-tidy with every bugprone-* check
promoted to an error, so this failed the build.

Not a real bug -- n/k/tileRows/tileCols are MLIR shape dimensions and
therefore never negative, so the round trip through uint64_t is always
value-preserving here -- but "obviously fine by inspection" is exactly
what an explicit static_cast should say instead of leaving implicit, per
the project's own .clang-tidy header ("selected for signal... a config
that fires on idiomatic MLIR trains everyone to ignore the output").

Verified: clang-tidy-18 no longer flags either line under
bugprone-narrowing-conversions; the full FileCheck/lit suite (29/29,
including the ragged-padding structural tests this code path is for),
both ctest suites, and the 78-test pytest suite all still pass from a
clean rebuild. No behavior change -- both padded dimensions are
identical before and after.

+++ lib/Transforms/CIMPartition.cpp
@@ -157,8 +157,18 @@ struct CIMPartitionPass : public CIMPartitionBase<CIMPartitionPass> {
     // paddedK == k and every branch below that only fires "when padded"
     // is simply dead, so the exact-multiple case emits identical IR to
     // before this padding support existed.
-    const int64_t paddedN = llvm::divideCeil(n, tileRows) * tileRows;
-    const int64_t paddedK = llvm::divideCeil(k, tileCols) * tileCols;
+    //
+    // llvm::divideCeil takes and returns uint64_t; n/k/tileRows/tileCols
+    // are shape dimensions and therefore never negative (MLIR shapes are
+    // non-negative by construction), so the explicit cast back to int64_t
+    // here is exactly the widen-then-narrow round trip that is always
+    // value-preserving for this input, not a silent truncation risk --
+    // spelled out rather than left implicit, which is what
+    // bugprone-narrowing-conversions is flagging.
+    const int64_t paddedN =
+        static_cast<int64_t>(llvm::divideCeil(n, tileRows) * tileRows);
+    const int64_t paddedK =
+        static_cast<int64_t>(llvm::divideCeil(k, tileCols) * tileCols);
     const bool needsPadding = paddedN != n || paddedK != k;
 
     OpBuilder builder(op);
